# doc/mall_admin/swagger.yaml
swagger: "2.0"
definitions:
  dto.AdminRegisterDTO:
    properties:
      email:
        description: 邮箱
        type: string
      icon:
        description: 图标
        type: string
      nickname:
        description: 昵称
        type: string
      note:
        description: 备注
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminRoleRelationDTO:
    properties:
      adminId:
        type: integer
      roleIds:
        items:
          type: integer
        type: array
    type: object
...
  dto.AdminUpdatePasswordDTO:
    properties:
      newPassword:
        description: 新密码
        type: string
      oldPassword:
        description: 旧密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - newPassword
    - oldPassword
    - username
    type: object
...
  dto.MenuCreateDTO:
    properties:
      hidden:
        description: 前端隐藏
        type: integer
      icon:
        description: 前端图标
        type: string
      level:
        description: 菜单级数
        type: integer
      name:
        description: 前端名称
        type: string
      parentId:
        description: 父级ID
        type: integer
      sort:
        description: 菜单排序
        type: integer
      title:
        description: 菜单名称
        type: string
    type: object
  dto.MenuUpdateHiddenDTO:
    properties:
      hidden:
        description: 前段隐藏菜单
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /admin/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除管理员
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 删除管理员
      tags:
      - 管理员
  /admin/info:
    get:
      consumes:
      - application/json
      description: 获取管理员信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 获取管理员信息
      tags:
      - 管理员
  /admin/list:
    get:
      consumes:
      - application/json
      description: 分页查询管理员列表
      parameters:
      - description: 分页查询参数
        in: query
        name: data
        required: true
        type: dto.PublicPageQuery
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 分页查询管理员列表
      tags:
      - 管理员
  /admin/register:
    post:
      consumes:
      - application/json
      description: 管理员注册
      parameters:
      - description: 管理员注册
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminRegisterDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 管理员注册
      tags:
      - 管理员
...
  /admin/role/update:
    post:
      consumes:
      - application/json
      description: 更新管理员信息
      parameters:
      - description: 管理员ID
        in: path
        name: id
        required: true
        type: integer
      - description: 管理员信息更新
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminRoleRelationDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 更新管理员信息
      tags:
      - 管理员
...
  /admin/updatePassword:
    post:
      consumes:
      - application/json
      description: 更新管理员密码
      parameters:
      - description: 管理员信息更新
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdatePasswordDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 更新管理员密码
      tags:
      - 管理员
...
  /menu/create:
    post:
      consumes:
      - application/json
      description: 菜单创建
      parameters:
      - description: 菜单创建
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.MenuCreateDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 菜单创建
      tags:
      - 菜单
  /menu/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 删除菜单
      tags:
      - 菜单
...
  /menu/treeList:
    get:
      consumes:
      - application/json
      description: 获取菜单树
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 获取菜单树
      tags:
      - 菜单
...
  /menu/updateHidden/{id}:
    post:
      consumes:
      - application/json
      description: 更新菜单显示状态
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateHiddenDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 更新菜单显示状态
      tags:
      - 菜单
...